FBManager/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
cryptography
pyyaml
python-dotenv
apscheduler
playwright
loguru
jinja2
python-multipart
